Revert inner class change to minimise diff
