+++ integrationtests/src/test/java/io/kroxylicious/proxy/OutOfBandRequestIT.java
@@ -22,6 +22,7 @@ import org.junit.jupiter.api.Test;
 import io.kroxylicious.proxy.config.FilterDefinition;
 import io.kroxylicious.proxy.config.FilterDefinitionBuilder;
 import io.kroxylicious.proxy.filter.OutOfBandSendFilterFactory;
+import io.kroxylicious.proxy.filter.RequestResponseMarkingFilter;
 import io.kroxylicious.proxy.filter.RequestResponseMarkingFilterFactory;
 import io.kroxylicious.test.Request;
 import io.kroxylicious.test.Response;
@@ -32,7 +33,7 @@ import io.kroxylicious.test.tester.KroxyliciousTesters;
 import io.kroxylicious.test.tester.MockServerKroxyliciousTester;
 
 import static io.kroxylicious.UnknownTaggedFields.unknownTaggedFieldsToStrings;
-import static io.kroxylicious.proxy.filter.RequestResponseMarkingFilterFactory.Filter.FILTER_NAME_TAG;
+import static io.kroxylicious.proxy.filter.RequestResponseMarkingFilter.FILTER_NAME_TAG;
 import static org.apache.kafka.common.protocol.ApiKeys.CREATE_TOPICS;
 import static org.apache.kafka.common.protocol.ApiKeys.DESCRIBE_CLUSTER;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -91,12 +92,12 @@ public class OutOfBandRequestIT {
         Request request = tester.getOnlyRequestForApiKey(CREATE_TOPICS);
         String tags = unknownTaggedFieldsToStrings(request.message(), FILTER_NAME_TAG)
                 .collect(Collectors.joining(","));
-        assertEquals(RequestResponseMarkingFilterFactory.class.getSimpleName() + ".Filter-upstreamOfOutOfBandFilter-request", tags);
+        assertEquals(RequestResponseMarkingFilter.class.getSimpleName() + "-upstreamOfOutOfBandFilter-request", tags);
     }
 
     private static void thenResponseContainsTagsAugmentedInByUpstreamFilterOnly(DescribeClusterResponseData responseData) {
-        assertEquals("filterNameTaggedFieldsFromOutOfBandResponse: " + RequestResponseMarkingFilterFactory.class.getSimpleName()
-                + ".Filter-upstreamOfOutOfBandFilter-response", responseData.errorMessage());
+        assertEquals("filterNameTaggedFieldsFromOutOfBandResponse: " + RequestResponseMarkingFilter.class.getSimpleName()
+                + "-upstreamOfOutOfBandFilter-response", responseData.errorMessage());
     }
 
     private static DescribeClusterResponseData whenDescribeCluster(KafkaClient client) {
