+++ kroxylicious-additional-filters/kroxylicious-schema-validation/src/test/java/io/kroxylicious/proxy/filter/schema/ProduceValidationFilterFactoryTest.java
@@ -32,7 +32,7 @@ class ProduceValidationFilterFactoryTest {
         ProduceValidationFilterFactory factory = new ProduceValidationFilterFactory();
         ValidationConfig config = new ValidationConfig(true, List.of(), new RecordValidationRule(null, null));
         Filter filter = factory.createFilter(null, config);
-        assertThat(filter).isNotNull().isInstanceOf(ProduceValidationFilterFactory.Filter.class);
+        assertThat(filter).isNotNull().isInstanceOf(ProduceValidationFilter.class);
     }
 
 }
