+++ kroxylicious-sample/src/test/java/io/kroxylicious/sample/SampleFetchResponseFilterFactoryFilterTest.java
@@ -63,14 +63,14 @@ class SampleFetchResponseFilterFactoryFilterTest {
 
     @Captor
     private ArgumentCaptor<ApiMessage> apiMessageCaptor;
-    private SampleFetchResponseFilterFactory.Filter filter;
+    private SampleFetchResponseFilter filter;
     private ResponseHeaderData headerData;
 
     @BeforeEach
     public void beforeEach() {
         setupContextMock();
         SampleFilterConfig config = new SampleFilterConfig(CONFIG_FIND_VALUE, CONFIG_REPLACE_VALUE);
-        this.filter = new SampleFetchResponseFilterFactory.Filter(config);
+        this.filter = new SampleFetchResponseFilter(config);
         this.headerData = new ResponseHeaderData();
     }
